odoo-pythonanywhere/requirements.txt
python-dotenv>=1.0.0
tqdm>=4.65.0
Flask>=3.0.0
